fix: corrected the typscript file
