+++ tsconfig.json
@@ -1,26 +1,15 @@
 {
   "compilerOptions": {
+    "declaration": true,
+    "strictNullChecks": true,
+    "target": "es5", 
+    "outDir": "dist",
     "module": "commonjs",
     "moduleResolution": "node",
-    "target": "es6",
     "sourceMap": true,
-    "removeComments": true,
-    "experimentalDecorators": true,
-    "emitDecoratorMetadata": true,
-    "isolatedModules": false,
-    "lib": [
-      "esnext",
-      "dom"
-    ],
-    "typeRoots": [
-      "node_modules/@types"
-    ]
+    "lib": ["es2015", "dom"],
+    "rootDir": "src"
   },
-
-  "exclude": [
-    "node_modules",
-    "typings",
-    "typings/globals",
-    "typings/index.d.ts"
-  ]
+  "include": ["src"],
+  "exclude": ["node_modules", "dist"]
 }
